+++ CHANGELOG.md
@@ -9,9 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Fix typos in payloads ([#57][pr57]):
+  - `get_updates`: `offset` `i64` -> `i32`
+  - `send_location`: make `live_period` optional
 - `send_contact` signature (`phone_number` and `first_name` `f64` => `String`) ([#56][pr56])
 
 [pr56]: https://github.com/teloxide/teloxide-core/pull/56
+[pr57]: https://github.com/teloxide/teloxide-core/pull/57
 
 ## [0.1.1] - 2020-02-17